Move download function to detail screen.
